feat(router): T8 rewire rsc-poc onto dual-mode 0.4.0-rc.3
Wire apps/rsc-poc to renderRequest + registered routes (GAP-194 T8).
Split SPA SSR (server-ssr) and Router core export so the RSC graph
never pulls react-dom/server or createContext.
