+++ apps/rsc-poc/src/entry.rsc.tsx
@@ -1,3 +1,9 @@
+/**
+ * RSC entry — T8 rewire onto `@revealui/router` `renderRequest` + registered routes.
+ * Pathname if/else shim removed; match + actions owned by the dual-mode router.
+ */
+
+import { renderRequest } from '@revealui/router/server';
 import {
   createTemporaryReferenceSet,
   decodeAction,
@@ -7,9 +13,9 @@ import {
   renderToReadableStream,
 } from '@vitejs/plugin-rsc/rsc';
 import type { ReactFormState } from 'react-dom/client';
-import { ActionsPage } from './pages/actions.tsx';
-import { CounterPage } from './pages/counter.tsx';
-import { HomePage } from './pages/home.tsx';
+import { createAppRouter } from './app-router.ts';
+import { AppLayout } from './pages/layout.tsx';
+import { NotFoundPage } from './pages/not-found.tsx';
 
 export interface RscPayload {
   root: React.ReactNode;
@@ -17,79 +23,35 @@ export interface RscPayload {
   formState?: ReactFormState;
 }
 
-function resolvePageComponent(pathname: string): React.ReactNode {
-  if (pathname === '/counter') {
-    return <CounterPage />;
-  }
-  if (pathname === '/actions') {
-    return <ActionsPage />;
-  }
-  return <HomePage />;
-}
+const router = createAppRouter();
 
-function NavBar(): React.ReactNode {
-  return (
-    <nav
-      style={{
-        padding: '12px 16px',
-        borderBottom: '1px solid #e5e7eb',
-        marginBottom: '24px',
-        display: 'flex',
-        gap: '16px',
-      }}
-    >
-      <a href="/">Home</a>
-      <a href="/counter">Counter</a>
-      <a href="/actions">Actions</a>
-    </nav>
-  );
-}
+function renderMatchedTree(match: ReturnType<typeof router.match>): React.ReactNode {
+  if (!match) {
+    return (
+      <AppLayout>
+        <NotFoundPage />
+      </AppLayout>
+    );
+  }
 
-function Layout({ children }: { children: React.ReactNode }): React.ReactNode {
-  return (
-    <html lang="en">
-      <head>
-        <meta charSet="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>RSC POC — Phase 2.1 spike</title>
-      </head>
-      <body>
-        <NavBar />
-        <main style={{ padding: '0 16px' }}>{children}</main>
-      </body>
-    </html>
-  );
+  const { route, params, data } = match;
+  const Page = route.component;
+  const Layout = route.layout ?? AppLayout;
+  const page = <Page params={params} data={data} />;
+  return <Layout>{page}</Layout>;
 }
 
 export default { fetch: handler };
 
 async function handler(request: Request): Promise<Response> {
-  const url = new URL(request.url);
-
-  let returnValue: RscPayload['returnValue'] | undefined;
   let formState: ReactFormState | undefined;
   let temporaryReferences: unknown | undefined;
-  let actionStatus: number | undefined;
 
+  // Progressive form POST (no x-rsc-action) stays in the consumer until 2.2.4
+  // progressive-enhancement is first-class on the router.
   const actionId = request.headers.get('x-rsc-action');
   const isFormAction = request.method === 'POST' && !actionId;
-
-  if (actionId) {
-    const contentType = request.headers.get('content-type');
-    const body = contentType?.startsWith('multipart/form-data')
-      ? await request.formData()
-      : await request.text();
-    temporaryReferences = createTemporaryReferenceSet();
-    const args = await decodeReply(body, { temporaryReferences });
-    const action = await loadServerAction(actionId);
-    try {
-      const data = await (action as (...a: unknown[]) => Promise<unknown>)(...(args as unknown[]));
-      returnValue = { ok: true, data };
-    } catch (e) {
-      returnValue = { ok: false, data: e };
-      actionStatus = 500;
-    }
-  } else if (isFormAction) {
+  if (isFormAction) {
     const formData = await request.formData();
     const decodedAction = await decodeAction(formData);
     try {
@@ -100,31 +62,43 @@ async function handler(request: Request): Promise<Response> {
     }
   }
 
-  const page = resolvePageComponent(url.pathname);
-  const rscPayload: RscPayload = {
-    root: <Layout>{page}</Layout>,
+  return renderRequest(request, {
+    router,
     formState,
-    returnValue,
-  };
-  const rscStream = renderToReadableStream<RscPayload>(rscPayload, { temporaryReferences });
-
-  const isRscRequest =
-    url.pathname.endsWith('.rsc') || request.headers.get('accept') === 'text/x-component';
-  if (isRscRequest) {
-    return new Response(rscStream, {
-      status: actionStatus,
-      headers: { 'content-type': 'text/x-component;charset=utf-8' },
-    });
-  }
-
-  const ssrEntry = await import.meta.viteRsc.loadModule<typeof import('./entry.ssr.tsx')>(
-    'ssr',
-    'index',
-  );
-  const htmlStream = await ssrEntry.renderHTML(rscStream, { formState });
-
-  return new Response(htmlStream, {
-    headers: { 'content-type': 'text/html' },
+    title: 'RSC POC — dual-mode router',
+    loadServerAction: async (id) => {
+      const action = await loadServerAction(id);
+      return action as (...args: unknown[]) => unknown | Promise<unknown>;
+    },
+    decodeActionArgs: async (req) => {
+      const contentType = req.headers.get('content-type');
+      const body = contentType?.startsWith('multipart/form-data')
+        ? await req.formData()
+        : await req.text();
+      temporaryReferences = createTemporaryReferenceSet();
+      const args = await decodeReply(body, { temporaryReferences });
+      return args as unknown[];
+    },
+    createRscStream: async (_req, ctx) => {
+      router.seedCurrentMatch(ctx.match);
+      const rscPayload: RscPayload = {
+        root: renderMatchedTree(ctx.match),
+        formState,
+        returnValue: ctx.returnValue,
+      };
+      return renderToReadableStream<RscPayload>(rscPayload, { temporaryReferences });
+    },
+    loadBootstrapScriptContent: () => import.meta.viteRsc.loadBootstrapScriptContent('index'),
+    renderHtml: async ({ rscStream, bootstrapScriptContent, formState: fs }) => {
+      const ssrEntry = await import.meta.viteRsc.loadModule<typeof import('./entry.ssr.tsx')>(
+        'ssr',
+        'index',
+      );
+      return ssrEntry.renderHTML(rscStream, {
+        formState: fs as ReactFormState | undefined,
+        bootstrapScriptContent,
+      });
+    },
   });
 }
