+++ apps/rsc-poc/src/entry.ssr.tsx
@@ -1,3 +1,8 @@
+/**
+ * SSR entry — consumes teed RSC flight half + bootstrap that already inlines
+ * `__RSC_PAYLOAD__` (via `@revealui/router/server` `inlineRscPayloadScript` in
+ * `renderRequest`). No second base64 pass (D15 / chunked encode stays in router).
+ */
 import { createFromReadableStream } from '@vitejs/plugin-rsc/ssr';
 import React from 'react';
 import type { ReactFormState } from 'react-dom/client';
@@ -6,40 +11,24 @@ import type { RscPayload } from './entry.rsc.tsx';
 
 export async function renderHTML(
   rscStream: ReadableStream<Uint8Array>,
-  options: { formState?: ReactFormState },
+  options: {
+    formState?: ReactFormState;
+    /** When set (from `renderRequest`), already includes payload script + client entry. */
+    bootstrapScriptContent?: string;
+  },
 ): Promise<ReadableStream<Uint8Array>> {
-  const [rscStream1, rscStream2] = rscStream.tee();
-
   let payload: Promise<RscPayload> | undefined;
   function SsrRoot(): React.ReactNode {
-    payload ??= createFromReadableStream<RscPayload>(rscStream1);
+    payload ??= createFromReadableStream<RscPayload>(rscStream);
     return React.use(payload).root;
   }
 
-  const bootstrapScriptContent = await import.meta.viteRsc.loadBootstrapScriptContent('index');
+  const bootstrapScriptContent =
+    options.bootstrapScriptContent ??
+    (await import.meta.viteRsc.loadBootstrapScriptContent('index'));
 
-  const chunks: Uint8Array[] = [];
-  const reader = rscStream2.getReader();
-  const _encoder = new TextEncoder();
-  while (true) {
-    const { done, value } = await reader.read();
-    if (done) break;
-    chunks.push(value);
-  }
-  const rscBytes = chunks.reduce((acc, c) => {
-    const merged = new Uint8Array(acc.length + c.length);
-    merged.set(acc);
-    merged.set(c, acc.length);
-    return merged;
-  }, new Uint8Array(0));
-  const rscBase64 = btoa(String.fromCharCode(...rscBytes));
-
-  const rscInlineScript = `self.__RSC_PAYLOAD__="${rscBase64}";`;
-
-  const htmlStream = await renderToReadableStream(<SsrRoot />, {
-    bootstrapScriptContent: rscInlineScript + bootstrapScriptContent,
+  return renderToReadableStream(<SsrRoot />, {
+    bootstrapScriptContent,
     formState: options.formState,
   });
-
-  return htmlStream;
 }
