+++ packages/router/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@revealui/router",
-  "version": "0.4.0-rc.2",
+  "version": "0.4.0-rc.3",
   "description": "Lightweight dual-mode router for React apps: client SPA (default) plus opt-in RSC mode. SSR, data loaders, middleware, nested layouts. Works with Vite, Hono, or any React setup.",
   "keywords": [
     "file-based-routing",
@@ -38,10 +38,20 @@
       "import": "./dist/index.js",
       "default": "./dist/index.js"
     },
+    "./core": {
+      "types": "./dist/core.d.ts",
+      "import": "./dist/core.js",
+      "default": "./dist/core.js"
+    },
     "./server": {
       "types": "./dist/server.d.ts",
       "import": "./dist/server.js",
       "default": "./dist/server.js"
+    },
+    "./server-ssr": {
+      "types": "./dist/server-ssr.d.ts",
+      "import": "./dist/server-ssr.js",
+      "default": "./dist/server-ssr.js"
     }
   },
   "files": [
