+++ packages/router/src/core.ts
@@ -1,0 +1,19 @@
+/**
+ * RSC-safe + client-safe core: Router class + types only.
+ * Does not import React context/hooks (those live in `./components`).
+ * Use this from dual-mode route tables shared by server and browser entries.
+ */
+export { Router } from './router';
+export type {
+  Location,
+  MiddlewareContext,
+  NavigateOptions,
+  Route,
+  RouteMatch,
+  RouteMeta,
+  RouteMiddleware,
+  RouteParams,
+  RouterMode,
+  RouterOptions,
+  RouterRscOptions,
+} from './types';
